move evaluation action to the organization rule controller
